Revert "refactor: 훅 위치 변경"
This reverts commit 5086bef45306f69ef465a95392404c431745bddf.
