+++ src/pages/MyPage.jsx
@@ -4,9 +4,9 @@ import { useNavigate } from 'react-router-dom'
 import { v4 as uuidv4 } from 'uuid'
 
 import Icon from '@/components/common/Icon'
+import { useGetPreSignedUrl } from '@/hooks/useAlbum'
 import { useLogout } from '@/hooks/useAuth'
 import { useUpdateUserInfo } from '@/hooks/useUser'
-import { useGetPreSignedUrl } from '@/queries/album/mutations'
 import axios from 'axios'
 import ConfirmModal from '../components/common/ConfirmModal'
 import Header from '../components/common/Header'
